// ui/lib/api/schemas/op.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 */
import type { Node } from './node'
import type { NodeId } from './nodeId'
import type { NodePatch } from './nodePatch'
import type { Page } from './page'
import type { PageId } from './pageId'
import type { PagePatch } from './pagePatch'
import type { ProjectMetaPatch } from './projectMetaPatch'

export type Op =
  | {
      updateProjectMeta: {
        patch: ProjectMetaPatch
        prev?: ProjectMetaPatch
      }
    }
  | {
      addPage: {
        /** @minimum 0 */
        at: number
        page: Page
      }
    }
  | {
      removePage: {
        id: PageId
        /** @minimum 0 */
        prev_index: number
        prev_page: Page
      }
    }
  | {
      updatePage: {
        id: PageId
        patch: PagePatch
        prev?: PagePatch
      }
    }
  | {
      reorderPages: {
        order: PageId[]
        prev_order: PageId[]
      }
    }
  | {
      addNode: {
        /** @minimum 0 */
        at: number
        node: Node
        page: PageId
      }
    }
  | {
      removeNode: {
        id: NodeId
        page: PageId
        /** @minimum 0 */
        prev_index: number
        prev_node: Node
      }
    }
  | {
      updateNode: {
        id: NodeId
        page: PageId
        patch: NodePatch
        prev?: NodePatch
      }
    }
  | {
      reorderNodes: {
        order: NodeId[]
        page: PageId
        prev_order: NodeId[]
      }
    }
  | {
      batch: {
        label: string
        ops: Op[]
      }
    }